requirements.txt
av==13.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
djitellopy==2.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
numpy==2.0.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
